+++ sdk/ml/azure-ai-ml/azure/ai/ml/_restclient/v2024_01_01_preview/aio/_azure_machine_learning_workspaces.py
@@ -16,7 +16,7 @@ from azure.mgmt.core import AsyncARMPipelineClient
 
 from .. import models
 from ._configuration import AzureMachineLearningWorkspacesConfiguration
-from .operations import BatchDeploymentsOperations, BatchEndpointsOperations, CapacityReservationGroupsOperations, CodeContainersOperations, CodeVersionsOperations, ComponentContainersOperations, ComponentVersionsOperations, ComputeOperations, DataContainersOperations, DataVersionsOperations, DatastoresOperations, EnvironmentContainersOperations, EnvironmentVersionsOperations, FeaturesOperations, FeaturesetContainersOperations, FeaturesetVersionsOperations, FeaturestoreEntityContainersOperations, FeaturestoreEntityVersionsOperations, InferenceEndpointsOperations, InferenceGroupsOperations, InferencePoolsOperations, JobsOperations, LabelingJobsOperations, ModelContainersOperations, ModelVersionsOperations, OnlineDeploymentsOperations, OnlineEndpointsOperations, QuotasOperations, RegistriesOperations, RegistryCodeContainersOperations, RegistryCodeVersionsOperations, RegistryComponentContainersOperations, RegistryComponentVersionsOperations, RegistryDataContainersOperations, RegistryDataVersionsOperations, RegistryEnvironmentContainersOperations, RegistryEnvironmentVersionsOperations, RegistryModelContainersOperations, RegistryModelVersionsOperations, SchedulesOperations, ServerlessEndpointsOperations, UsagesOperations, VirtualMachineSizesOperations, WorkspaceFeaturesOperations
+from .operations import BatchDeploymentsOperations, BatchEndpointsOperations, CapacityReservationGroupsOperations, CodeContainersOperations, CodeVersionsOperations, ComponentContainersOperations, ComponentVersionsOperations, ComputeOperations, DataContainersOperations, DataVersionsOperations, DatastoresOperations, EnvironmentContainersOperations, EnvironmentVersionsOperations, FeaturesOperations, FeaturesetContainersOperations, FeaturesetVersionsOperations, FeaturestoreEntityContainersOperations, FeaturestoreEntityVersionsOperations, InferenceEndpointsOperations, InferenceGroupsOperations, InferencePoolsOperations, JobsOperations, LabelingJobsOperations, MarketplaceSubscriptionsOperations, ModelContainersOperations, ModelVersionsOperations, OnlineDeploymentsOperations, OnlineEndpointsOperations, QuotasOperations, RegistriesOperations, RegistryCodeContainersOperations, RegistryCodeVersionsOperations, RegistryComponentContainersOperations, RegistryComponentVersionsOperations, RegistryDataContainersOperations, RegistryDataReferencesOperations, RegistryDataVersionsOperations, RegistryEnvironmentContainersOperations, RegistryEnvironmentVersionsOperations, RegistryModelContainersOperations, RegistryModelVersionsOperations, SchedulesOperations, ServerlessEndpointsOperations, UsagesOperations, VirtualMachineSizesOperations, WorkspaceFeaturesOperations
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
@@ -60,12 +60,18 @@ class AzureMachineLearningWorkspaces:    # pylint: disable=too-many-instance-att
     :ivar registry_data_versions: RegistryDataVersionsOperations operations
     :vartype registry_data_versions:
      azure.mgmt.machinelearningservices.aio.operations.RegistryDataVersionsOperations
+    :ivar registry_data_references: RegistryDataReferencesOperations operations
+    :vartype registry_data_references:
+     azure.mgmt.machinelearningservices.aio.operations.RegistryDataReferencesOperations
     :ivar registry_environment_containers: RegistryEnvironmentContainersOperations operations
     :vartype registry_environment_containers:
      azure.mgmt.machinelearningservices.aio.operations.RegistryEnvironmentContainersOperations
     :ivar registry_environment_versions: RegistryEnvironmentVersionsOperations operations
     :vartype registry_environment_versions:
      azure.mgmt.machinelearningservices.aio.operations.RegistryEnvironmentVersionsOperations
+    :ivar marketplace_subscriptions: MarketplaceSubscriptionsOperations operations
+    :vartype marketplace_subscriptions:
+     azure.mgmt.machinelearningservices.aio.operations.MarketplaceSubscriptionsOperations
     :ivar registry_model_containers: RegistryModelContainersOperations operations
     :vartype registry_model_containers:
      azure.mgmt.machinelearningservices.aio.operations.RegistryModelContainersOperations
@@ -189,8 +195,10 @@ class AzureMachineLearningWorkspaces:    # pylint: disable=too-many-instance-att
         self.registry_component_versions = RegistryComponentVersionsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.registry_data_containers = RegistryDataContainersOperations(self._client, self._config, self._serialize, self._deserialize)
         self.registry_data_versions = RegistryDataVersionsOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.registry_data_references = RegistryDataReferencesOperations(self._client, self._config, self._serialize, self._deserialize)
         self.registry_environment_containers = RegistryEnvironmentContainersOperations(self._client, self._config, self._serialize, self._deserialize)
         self.registry_environment_versions = RegistryEnvironmentVersionsOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.marketplace_subscriptions = MarketplaceSubscriptionsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.registry_model_containers = RegistryModelContainersOperations(self._client, self._config, self._serialize, self._deserialize)
         self.registry_model_versions = RegistryModelVersionsOperations(self._client, self._config, self._serialize, self._deserialize)
         self.batch_endpoints = BatchEndpointsOperations(self._client, self._config, self._serialize, self._deserialize)