+++ sdk/ai/azure-ai-resources/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 1.0.0b6 (2024-02-06)
+
+### Other Changes
+
+- Bug fixes
+
+## 1.0.0b5 (2024-02-01)
+
+### Other Changes
+
+- Duplicate cleanup
+
 ## 1.0.0b4 (2024-02-01)
 
 ### Other Changes